==> 02.Scripts/1.Preprocessing.ipynb <==
# QSR Data Challenge

## Preprocessing

Imports

import os
import pandas as pd
import datetime
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler

Get the current working directory:

def get_parent_dir(directory):
    import os
    return os.path.dirname(directory)

current_dirs_parent = get_parent_dir(os.getcwd())
dataraw_dir=current_dirs_parent+"/01.Data/Raw/"
preproc_dir=current_dirs_parent+"/01.Data/Preprocessing/"

Load the data:

data_a=pd.read_excel(dataraw_dir+"processminer-rare-event-detection-data-augmentation.xlsx", 
                     sheet_name='data-(a)-raw-data')
data_b=pd.read_excel(dataraw_dir+"processminer-rare-event-detection-data-augmentation.xlsx", 
                     sheet_name='data-(b)-4-min-ahead-conse-rmvd')
data_b.rename(columns={"y-4min-ahead":"y_lead2"},inplace=True)

Check first lines:

data_a.head()

data_b.head()

data_b["y_lead2"].sum()

data_b.shape

Define functions for data transformations:

#Add EventId
def event_id(df,yname="y"):
    dfout=df
    dfout["EventID"]=dfout[yname].shift(periods=-1,fill_value=0)
    dfout["EventID"]=dfout["EventID"].cumsum()+1    
    return(dfout)

#Cycles since last failure
def add_cycles(df):
    dfout=df
    dfout["Cycle"]=1
    dfout["Cycle"]=dfout.groupby("EventID")["Cycle"].cumsum()
    return(dfout)
    

#Lagged difference. 
##i_list: variable number as list, assuming are all named x
##n_list: number of lag periods as list
##df: pandas dataframe to add variables
def lagdif_xi_n(df,i_list=[1,2,3],n_list=[1,2,3]):
    dfout=df
    for n in n_list:
        for i in i_list:
            dfout["x"+str(i)+"_df_l"+str(n)]=dfout["x"+str(i)].diff(periods=n)
    
    return(dfout)

#Second order Lagged difference. 
##i: variable number as list, assuming are all named x 
##n: number of lag periods as list
##df: pandas dataframe to add variables

def lagdif2_xi_n(df,i_list=[1,2,3],n_list=[1,2,3]):
    dfout=df
    for n in n_list:
        for i in i_list:
            dfout["x"+str(i)+"_df2_l"+str(n)]=dfout["x"+str(i)].diff(periods=n).diff(periods=n)
            
    return(dfout)
            

#Lagged percentual difference. 
##i_list: variable number as list, assuming are all named x
##n_list: number of lag periods as list
##df: pandas dataframe to add variables

def perc_lagdif_xi_n(df,i_list=[1,2,3],n_list=[1,2,3]):
    dfout=df
    for n in n_list:
        for i in i_list:
            dfout["x"+str(i)+"_pdf_l"+str(n)]=dfout["x"+str(i)].pct_change(periods=n)
    
    return(dfout)

#Lagged absolute percentual difference
##i_list: variable number as list, assuming are all named x
##n_list: number of lag periods as list
##df: pandas dataframe to add variables

def abs_perc_lagdif_xi_n(df,i_list=[1,2,3],n_list=[1,2,3]):
    dfout=df
    for n in n_list:
        for i in i_list:
            dfout["x"+str(i)+"_apdf_l"+str(n)]=dfout["x"+str(i)].pct_change(periods=n).abs()
            
    return(dfout)

#Exponential moving average percentual difference
##i_list: variable number as list, assuming are all named x
##n_list: number of lag periods as list
##alpha_list: smoothing parameter as list
##df: pandas dataframe to add variables

def EMW_perc_lagdif_xi_n(df,i_list=[1,2,3],n_list=[1,2,3],alpha_list=[0.25,0.5,0.75]):
    dfout=df
    for a in alpha_list:
        for n in n_list:        
            for i in i_list:
                dfout["x"+str(i)+"_emwpdf_a"+str(a)+"_l"+str(n)]=dfout["x"+str(i)].pct_change(periods=n).ewm(alpha=a).mean()
            
    return(dfout)

#Lagged log percentual difference 
##i_list: variable number as list, assuming are all named x
##n_list: number of lag periods as list
##df: pandas dataframe to add variables

def log_perc_lagdif_xi_n(df,i_list=[1,2,3],n_list=[1,2,3]):
    dfout=df
    for n in n_list:
        for i in i_list:
            dfout["x"+str(i)+"_logpdf_l"+str(n)]=np.log(dfout["x"+str(i)].pct_change(periods=n)+1)
    
    return(dfout)

#Exponential moving average of log percentual difference
##i_list: variable number as list, assuming are all named x
##n_list: number of lag periods as list
##alpha_list: smoothing parameter as list
##df: pandas dataframe to add variables

def EMW_log_perc_lagdif_xi_n(df,i_list=[1,2,3],n_list=[1,2,3],alpha_list=[0.25,0.5,0.75]):
    dfout=df
    for a in alpha_list:
        for n in n_list:        
            for i in i_list:
                dfout["x"+str(i)+"_emwpdf_a"+str(a)+"_l"+str(n)]=_
                np.log(dfout["x"+str(i)].pct_change(periods=n)+1).ewm(alpha=a).mean()
            
    return(dfout)

Date functions:

#Extracting:
##xday
def add_day(df):
    dfout=df    
    dfout["xday"]=dfout["time"].dt.day
    return(dfout)

##xhour
def add_hour(df):
    dfout=df    
    dfout["xhour"]=dfout["time"].dt.hour
    return(dfout)

##xminute
def add_minute(df):
    dfout=df    
    dfout["xminute"]=dfout["time"].dt.minute
    return(dfout)

##Month
def add_month(df):
    dfout=df    
    dfout["xmonth"]=dfout["time"].dt.month
    return(dfout)

#Turn into dummies
def date_dummies(df,varlist=["xday","xmonth","xhour","xminute"]):
    dfout=df
    for var in varlist:
        dfout=pd.concat([dfout,pd.get_dummies(dfout[var],prefix=var)],axis=1)
    return(dfout)

#Time difference from each row in minutes... minus 2 minutes
def time_diff(df):
    dfout=df
    dfout["xtimedif"]=dfout["time"].diff(periods=1)/np.timedelta64(1,'m')-2
    return(dfout)

#Dummy if there was a skip
def add_skip(df):
    dfout=df
    dfout["xskip"]=[1 if x > 0 else 0 for x in df['xtimedif']]
    return(dfout)


#SkipNumber
def add_skipnumber(df):
    dfout=df
    dfout["xskipid"]=dfout["xskip"].cumsum()+1
    return(dfout)

#Time since last skip
def add_sinceskip(df):
    dfout=df
    dfout["xsinceskip"]=1
    dfout["xsinceskip"]=dfout.groupby("xskipid")["xsinceskip"].cumsum()
    return(dfout)

Separate training and test data: 10% at the end for testing

train_b, test_b = train_test_split(data_b,test_size=0.1,shuffle=False)

train_b=train_b.copy()
test_b=test_b.copy()

first_test_date=test_b["time"].iloc[0]

#Data a should be split according to the date on data_b test set:
train_a = data_a.loc[data_a["time"] < first_test_date]
test_a = data_a.loc[data_a["time"] >= first_test_date]

train_a=train_a.copy()
test_a=test_a.copy()

train_b.tail()

test_b.head()

train_b["y_lead2"].sum()

test_b["y_lead2"].sum()

train_b.shape

test_b.shape

Pipeline to add new variables:

def preprocessing_pipeline(df):
    dfout=df.copy()
    dfout=lagdif_xi_n(df=dfout,i_list=range(1,62),n_list=range(4))
    dfout=perc_lagdif_xi_n(df=dfout,i_list=range(1,62),n_list=range(4))
    dfout=abs_perc_lagdif_xi_n(df=dfout,i_list=range(1,62),n_list=range(4))
    dfout=EMW_perc_lagdif_xi_n(df=dfout,i_list=range(1,61),n_list=range(4),alpha_list=[0.1,0.5,0.95])
    dfout=lagdif2_xi_n(df=dfout,i_list=range(1,62),n_list=range(4))
    
    #dfout=log_perc_lagdif_xi_n(df=dfout,i_list=range(1,61),n_list=range(4))
    #dfout=EMW_log_perc_lagdif_xi_n(df=dfout,i_list=range(1,61),n_list=range(4),alpha_list=[0.1,0.5,0.95])

    dfout=add_minute(dfout)
    dfout=add_hour(dfout)
    dfout=add_day(dfout)
    dfout=add_month(dfout)

    dfout=time_diff(dfout)
    dfout=add_skip(dfout)
    dfout=add_skipnumber(dfout)
    dfout=add_sinceskip(dfout)

    return(dfout)

Pipeline for data_b:

train_b_preproc=None
test_b_preproc=None
train_b_preproc=preprocessing_pipeline(train_b)
test_b_preproc=preprocessing_pipeline(test_b)

train_b_preproc.tail()

test_b_preproc.head()

Write to file

train_b_preproc.to_csv(preproc_dir+"train_b_preproc1.csv",index=False)
test_b_preproc.to_csv(preproc_dir+"test_b_preproc1.csv",index=False)

test_b_preproc["y_lead2"].sum()

train_b_preproc["y_lead2"].sum()

test_b_preproc.shape

train_b_preproc.shape

Pipeline for data a:

train_a_preproc=None
test_a_preproc=None
train_a_preproc=preprocessing_pipeline(train_a)
test_a_preproc=preprocessing_pipeline(test_a)

##Do this only before adding to train set b
#train_a_preproc=train_a_preproc.add_suffix("_aug")
#test_a_preproc=test_a_preproc.add_suffix("_aug")

train_a_preproc.tail()

test_a_preproc.head()

Write to file

train_a_preproc.to_csv(preproc_dir+"train_a_preproc1.csv",index=False)
test_a_preproc.to_csv(preproc_dir+"test_a_preproc1.csv",index=False)

Scale first:

train_b.head()

#Select X
selcols=[a for a in train_b if a.startswith("x")]
X_train_b=train_b[train_b.columns.intersection(selcols)].copy()
X_test_b=test_b[test_b.columns.intersection(selcols)].copy()

#X_train_b.fillna(0, inplace=True) 
#X_train_b.replace(to_replace=np.inf, value=0, inplace=True)

#X_test_b.fillna(0, inplace=True) 
#X_test_b.replace(to_replace=np.inf, value=0, inplace=True)

scaler= StandardScaler().fit(X_train_b)
X_train_b_sc=scaler.transform(X_train_b)
X_test_b_sc=scaler.transform(X_test_b)

unselcols=[a for a in train_b if not(a.startswith("x"))]
un_train_b=train_b[train_b.columns.intersection(unselcols)].copy()
un_test_b=test_b[test_b.columns.intersection(unselcols)].copy()
un_test_b.reset_index(inplace=True)

un_test_b.head()

X_train_b_sc_df=pd.DataFrame(columns=selcols,data=X_train_b_sc)
X_test_b_sc_df=pd.DataFrame(columns=selcols,data=X_test_b_sc)

X_train_b_sc_df.tail()

X_test_b_sc_df.head()

train_b_sc=pd.concat([un_train_b,X_train_b_sc_df],axis=1)
test_b_sc=pd.concat([un_test_b,X_test_b_sc_df],axis=1)

train_b_sc.tail()

test_b_sc.head()

Preprocess second:

train_b_preproc_sc=None
test_b_preproc_sc=None
train_b_preproc_sc=preprocessing_pipeline(train_b_sc)
test_b_preproc_sc=preprocessing_pipeline(test_b_sc)

train_b_preproc_sc.tail()

test_b_preproc_sc.head()

train_b_preproc_sc.to_csv(preproc_dir+"train_b_preproc_sc.csv",index=False)
test_b_preproc_sc.to_csv(preproc_dir+"test_b_preproc_sc.csv",index=False)
