==> 02.Scripts/5. Prediction on New Data.ipynb <==
# 5. Prediction on New Data

Assuming that the new data is in the folder "01.Data/New Data/" and is an excel file named "NewData.xlsx" with a worksheet named "NewData"

Imports:

import os
import pandas as pd
import pickle
import numpy as np
import xgboost as xgb
import sklearn.metrics as skmetric

Directories:

def get_parent_dir(directory):
    import os
    return os.path.dirname(directory)

current_dirs_parent = get_parent_dir(os.getcwd())
dataraw_dir=current_dirs_parent+"/01.Data/Raw/"
preproc_dir=current_dirs_parent+"/01.Data/Preprocessing/"
models_a_dir=current_dirs_parent+"/03.Models/Set A/"
models_b_dir=current_dirs_parent+"/03.Models/Set B/"
datanew_dir=current_dirs_parent+"/01.Data/New Data/"

Read data:

newdata=pd.read_excel(datanew_dir+"NewData.xlsx", sheet_name='NewData')

newdata.head()

Transformations:

#Add EventId
def event_id(df,yname="y"):
    dfout=df
    dfout["EventID"]=dfout[yname].shift(periods=-1,fill_value=0)
    dfout["EventID"]=dfout["EventID"].cumsum()+1    
    return(dfout)

#Cycles since last failure
def add_cycles(df):
    dfout=df
    dfout["Cycle"]=1
    dfout["Cycle"]=dfout.groupby("EventID")["Cycle"].cumsum()
    return(dfout)
    

#Lagged difference. 
##i_list: variable number as list, assuming are all named x
##n_list: number of lag periods as list
##df: pandas dataframe to add variables
def lagdif_xi_n(df,i_list=[1,2,3],n_list=[1,2,3]):
    dfout=df
    for n in n_list:
        for i in i_list:
            dfout["x"+str(i)+"_df_l"+str(n)]=dfout["x"+str(i)].diff(periods=n)
    
    return(dfout)

#Second order Lagged difference. 
##i: variable number as list, assuming are all named x 
##n: number of lag periods as list
##df: pandas dataframe to add variables

def lagdif2_xi_n(df,i_list=[1,2,3],n_list=[1,2,3]):
    dfout=df
    for n in n_list:
        for i in i_list:
            dfout["x"+str(i)+"_df2_l"+str(n)]=dfout["x"+str(i)].diff(periods=n).diff(periods=n)
            
    return(dfout)
            

#Lagged percentual difference. 
##i_list: variable number as list, assuming are all named x
##n_list: number of lag periods as list
##df: pandas dataframe to add variables

def perc_lagdif_xi_n(df,i_list=[1,2,3],n_list=[1,2,3]):
    dfout=df
    for n in n_list:
        for i in i_list:
            dfout["x"+str(i)+"_pdf_l"+str(n)]=dfout["x"+str(i)].pct_change(periods=n)
    
    return(dfout)

#Lagged absolute percentual difference
##i_list: variable number as list, assuming are all named x
##n_list: number of lag periods as list
##df: pandas dataframe to add variables

def abs_perc_lagdif_xi_n(df,i_list=[1,2,3],n_list=[1,2,3]):
    dfout=df
    for n in n_list:
        for i in i_list:
            dfout["x"+str(i)+"_apdf_l"+str(n)]=dfout["x"+str(i)].pct_change(periods=n).abs()
            
    return(dfout)

#Exponential moving average percentual difference
##i_list: variable number as list, assuming are all named x
##n_list: number of lag periods as list
##alpha_list: smoothing parameter as list
##df: pandas dataframe to add variables

def EMW_perc_lagdif_xi_n(df,i_list=[1,2,3],n_list=[1,2,3],alpha_list=[0.25,0.5,0.75]):
    dfout=df
    for a in alpha_list:
        for n in n_list:        
            for i in i_list:
                dfout["x"+str(i)+"_emwpdf_a"+str(a)+"_l"+str(n)]=dfout["x"+str(i)].pct_change(periods=n).ewm(alpha=a).mean()
            
    return(dfout)

#Lagged log percentual difference 
##i_list: variable number as list, assuming are all named x
##n_list: number of lag periods as list
##df: pandas dataframe to add variables

def log_perc_lagdif_xi_n(df,i_list=[1,2,3],n_list=[1,2,3]):
    dfout=df
    for n in n_list:
        for i in i_list:
            dfout["x"+str(i)+"_logpdf_l"+str(n)]=np.log(dfout["x"+str(i)].pct_change(periods=n)+1)
    
    return(dfout)

#Exponential moving average of log percentual difference
##i_list: variable number as list, assuming are all named x
##n_list: number of lag periods as list
##alpha_list: smoothing parameter as list
##df: pandas dataframe to add variables

def EMW_log_perc_lagdif_xi_n(df,i_list=[1,2,3],n_list=[1,2,3],alpha_list=[0.25,0.5,0.75]):
    dfout=df
    for a in alpha_list:
        for n in n_list:        
            for i in i_list:
                dfout["x"+str(i)+"_emwpdf_a"+str(a)+"_l"+str(n)]=_
                np.log(dfout["x"+str(i)].pct_change(periods=n)+1).ewm(alpha=a).mean()
            
    return(dfout)

Date functions:

#Extracting:
##xday
def add_day(df):
    dfout=df    
    dfout["xday"]=dfout["time"].dt.day
    return(dfout)

##xhour
def add_hour(df):
    dfout=df    
    dfout["xhour"]=dfout["time"].dt.hour
    return(dfout)

##xminute
def add_minute(df):
    dfout=df    
    dfout["xminute"]=dfout["time"].dt.minute
    return(dfout)

##Month
def add_month(df):
    dfout=df    
    dfout["xmonth"]=dfout["time"].dt.month
    return(dfout)

#Turn into dummies
def date_dummies(df,varlist=["xday","xmonth","xhour","xminute"]):
    dfout=df
    for var in varlist:
        dfout=pd.concat([dfout,pd.get_dummies(dfout[var],prefix=var)],axis=1)
    return(dfout)

#Time difference from each row in minutes... minus 2 minutes
def time_diff(df):
    dfout=df
    dfout["xtimedif"]=dfout["time"].diff(periods=1)/np.timedelta64(1,'m')-2
    return(dfout)

#Dummy if there was a skip
def add_skip(df):
    dfout=df
    dfout["xskip"]=[1 if x > 0 else 0 for x in df['xtimedif']]
    return(dfout)


#SkipNumber
def add_skipnumber(df):
    dfout=df
    dfout["xskipid"]=dfout["xskip"].cumsum()+1
    return(dfout)

#Time since last skip
def add_sinceskip(df):
    dfout=df
    dfout["xsinceskip"]=1
    dfout["xsinceskip"]=dfout.groupby("xskipid")["xsinceskip"].cumsum()
    return(dfout)

Pipeline:

def preprocessing_pipeline(df):
    dfout=df.copy()
    dfout=lagdif_xi_n(df=dfout,i_list=range(1,62),n_list=range(4))
    dfout=perc_lagdif_xi_n(df=dfout,i_list=range(1,62),n_list=range(4))
    dfout=abs_perc_lagdif_xi_n(df=dfout,i_list=range(1,62),n_list=range(4))
    dfout=EMW_perc_lagdif_xi_n(df=dfout,i_list=range(1,61),n_list=range(4),alpha_list=[0.1,0.5,0.95])
    dfout=lagdif2_xi_n(df=dfout,i_list=range(1,62),n_list=range(4))
    
    #dfout=log_perc_lagdif_xi_n(df=dfout,i_list=range(1,61),n_list=range(4))
    #dfout=EMW_log_perc_lagdif_xi_n(df=dfout,i_list=range(1,61),n_list=range(4),alpha_list=[0.1,0.5,0.95])

    dfout=add_minute(dfout)
    dfout=add_hour(dfout)
    dfout=add_day(dfout)
    dfout=add_month(dfout)

    dfout=time_diff(dfout)
    dfout=add_skip(dfout)
    dfout=add_skipnumber(dfout)
    dfout=add_sinceskip(dfout)

    return(dfout)

newdata_preproc=None
newdata_preproc=preprocessing_pipeline(newdata)

newdata_preproc.head()

selcols=[a for a in newdata_preproc.columns if a.startswith("x")]
X_newdata=newdata_preproc[newdata_preproc.columns.intersection(selcols)]
X_newdata=X_newdata.copy()
X_newdata.fillna(0, inplace=True) 
X_newdata.replace(to_replace=np.inf, value=0, inplace=True)

X_newdata.head()

Scaling this preprocessed data:

#Load scaler
scaler=pickle.load(open(models_b_dir+"clust_scaler.pickle.dat", "rb"))

X_newdata_sc=scaler.transform(X_newdata)

Load all clustering methods:

clf=pickle.load(open(models_b_dir+"clust_svdd.pickle.dat", "rb"))
db=pickle.load(open(models_b_dir+"clust_db.pickle.dat", "rb"))
isoforest=pickle.load(open(models_b_dir+"clust_isoforest.pickle.dat", "rb"))
pca=pickle.load(open(models_b_dir+"clust_pca.pickle.dat", "rb"))
fica=pickle.load(open(models_b_dir+"clust_fica.pickle.dat", "rb"))
svd=pickle.load(open(models_b_dir+"clust_svd.pickle.dat", "rb"))
lof=pickle.load(open(models_b_dir+"clust_lof.pickle.dat", "rb"))
tsnes=pickle.load(open(models_b_dir+"clust_tsnes.pickle.dat", "rb"))

Cluster the new data:

svdd_test=pd.DataFrame(clf.predict(X_newdata_sc))
dbscan_test=pd.DataFrame(db.fit_predict(X_newdata_sc))
isoforest_test=pd.DataFrame(isoforest.predict(X_newdata_sc))

test_pca_df=pd.DataFrame(pca.transform(X_newdata_sc))
test_fica_df=pd.DataFrame(fica.transform(X_newdata_sc))
test_svd_df=pd.DataFrame(svd.transform(X_newdata_sc))
test_tsne_df=pd.DataFrame(tsnes.fit_transform(X_newdata_sc))

test_lof_df=pd.DataFrame(lof.predict(X_newdata_sc))

Join in a new dataframe:

c_colnames=[*["x_svdd_"+str(i) for i in svdd_test.columns],\
          *["x_db_"+str(i) for i in dbscan_test.columns],\
          *["x_iso_"+str(i) for i in isoforest_test.columns],\
          *["x_pca_"+str(i) for i in test_pca_df.columns],\
          *["x_svd_"+str(i) for i in test_svd_df.columns],\
          *["x_lof_"+str(i) for i in test_lof_df],\
          *["x_tsne_"+str(i) for i in test_tsne_df.columns]]

clust_test_df=pd.concat([svdd_test,dbscan_test,isoforest_test,test_pca_df,test_svd_df,test_lof_df,test_tsne_df],axis=1)
clust_test_df.columns=c_colnames

clust_test_df.head()

Load all tree methods:

model2_blead2=pickle.load(open(models_b_dir+"xgboost_y_blead2.pickle.dat", "rb"))
model2_lead1=pickle.load(open(models_a_dir+"xgboost_y_lead1.pickle.dat", "rb"))
model2=pickle.load(open(models_a_dir+"xgboost_y.pickle.dat", "rb"))

dtest_b=xgb.DMatrix(X_newdata_sc)

y_lead2_pred_test_b_own = model2_blead2.predict(dtest_b)
y_pred_test_b = model2.predict(dtest_b)
y_lead1_pred_test_b = model2_lead1.predict(dtest_b)

xgb_test_b_own=pd.DataFrame({"xgb_y_prob" : y_pred_test_b,
                                 "xgb_ylead1_prob" : y_lead1_pred_test_b,
                                 #"xgb_ylead2_prob" : y_lead2_pred_test_b,
                                 "xgb_yblead2_prob" : y_lead2_pred_test_b_own})

xgb_test_b_own.head()

def logisticf(x):
    p=1/(1+np.exp(x))
    return p

logisticf(-0.167528)

xgb_test_b_prob=logisticf(xgb_test_b_own)
xgb_test_b_prob=xgb_test_b_prob.add_suffix("_prob")

xgb_test_b_prob.head()

Add moving average to XGB features:

def EMW_prob_lagdif_xi_n(df,n_list=[1,2,3],alpha_list=[0.25,0.5,0.75]):
    dfout=df.copy()
    i_list=dfout.columns
    for a in alpha_list:    
        for i in i_list:
            dfout[str(i)+"_emwpdf_a"+str(a)]=dfout[i].ewm(alpha=a).mean()            
    return(dfout)

xgb_test_b_emw=EMW_prob_lagdif_xi_n(xgb_test_b_own,alpha_list=[0.1,0.5,0.95])
xgb_test_b_prob_emw=EMW_prob_lagdif_xi_n(xgb_test_b_prob,alpha_list=[0.1,0.5,0.95])

xgb_test_b_emw.head()

xgb_test_b_prob_emw.head()

Concatenating all data before Logistic Regression:

X_test_sc_df=pd.DataFrame(columns=X_newdata.columns,data=X_newdata_sc)

X_test_sc_df.shape

clust_test_df.shape

xgb_test_b_emw.shape

X_test_sc_df.head()

final_b_test=pd.concat([X_test_sc_df,clust_test_df,xgb_test_b_emw,xgb_test_b_prob_emw],axis=1)

final_b_test.head()

LR13=pickle.load(open(models_b_dir+"LR13.pickle.dat", "rb"))

LR13

y_pred=LR13.predict(final_b_test)

#save as csv
pd.DataFrame({"y_pred":y_pred}).to_csv(datanew_dir+"y_pred.csv",index=False)
